Moved most pages back into root pages folder
