+++ src/pages/Details.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react'
 import { useParams } from 'react-router-dom'
-import { useAppDispatch, useAppSelector } from '../../hooks/redux'
+import { useAppDispatch, useAppSelector } from '../hooks/redux'
 
 export default function Details() {
   const { id } = useParams()
