+++ src/pages/Login.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom'
-import LoginForm from '../../components/LoginForm'
+import LoginForm from '../components/LoginForm'
 
 export default function Login() {
   const navigate = useNavigate()
