+++ src/pages/Register.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom'
-import RegisterForm from '../../components/RegisterForm'
+import RegisterForm from '../components/RegisterForm'
 
 export default function Register() {
   const navigate = useNavigate()
