+++ src/pages/Saved.tsx
@@ -1,4 +1,4 @@
-import ItemList from '../../components/ItemList'
+import ItemList from '../components/ItemList'
 
 export default function Saved() {
   return (
